+++ crates/backend-uzu/src/encodable_block/mixer/mamba2.rs
@@ -210,13 +210,13 @@ impl<B: Backend> Mixer<B> for Mamba2<B> {
         false
     }
 
-    fn max_context_length(&self) -> Option<usize> {
+    fn max_context_length(&self) -> Option<u32> {
         None
     }
 
     fn create_empty_state(
         &self,
-        _max_context_length: Option<usize>,
+        _max_context_length: Option<u32>,
         context: &B::Context,
     ) -> Result<Box<dyn MixerState<B>>, B::Error> {
         let mut conv_state = context.create_allocation(
