+++ crates/backend-uzu/src/engine/language_model/state.rs
@@ -12,7 +12,7 @@ pub struct LanguageModelState<B: Backend> {
     pub(super) prng: PRng,
     pub(super) transformer_state: TransformerState<B>,
     pub(super) speculator_state: Option<DFlashState<B>>,
-    pub(super) max_context_length: Option<usize>,
+    pub(super) max_context_length: Option<u32>,
 }
 
 impl<B: Backend> LanguageModelState<B> {
@@ -30,7 +30,7 @@ pub enum LanguageModelCreateEmptyStateError<B: Backend> {
 impl<B: Backend> LanguageModel<B> {
     pub fn create_empty_state(
         &self,
-        max_context_length: Option<usize>,
+        max_context_length: Option<u32>,
     ) -> Result<LanguageModelState<B>, LanguageModelCreateEmptyStateError<B>> {
         let tokens = Vec::new();
         let last_output_token = None;
@@ -46,9 +46,7 @@ impl<B: Backend> LanguageModel<B> {
             .speculator
             .as_ref()
             .map(|speculator| {
-                speculator.empty_state(
-                    (max_context_length.expect("speculator doesn't support unlimited state capacity")) as u32,
-                )
+                speculator.empty_state(max_context_length.expect("speculator doesn't support unlimited state capacity"))
             })
             .transpose()
             .map_err(LanguageModelCreateEmptyStateError::Backend)?;
